Refactoring of terms text update handling
- moved the creation of NSMutableAttributedString in a helper class
- added a static String var, that uses the same NSMutableAttributedString as displayed on the screen, to use for hash comparison
- renamed termsHash to acceptedTermsHash
