+++ ConcordiumWallet/Views/Login/TermsAndConditionsView/TermsAndConditionsPresenter.swift
@@ -34,7 +34,7 @@ class TermsAndConditionsPresenter {
 extension TermsAndConditionsPresenter: TermsAndConditionsPresenterProtocol {
     func userTappedAcceptTerms() {
         //save the hash of the accepted terms
-        AppSettings.termsHash = HashingHelper.hash(AppSettings.currentTerms)
+        AppSettings.acceptedTermsHash = HashingHelper.hash(TermsHelper.currentTerms)
         self.delegate?.userTappedAcceptTerms()
     }
 }
